// graphql/schema/generated/models/user.model.ts
/* Do not edit this file. It was generated programmatically. */

import { DEFAULT_MODEL_OPTIONS } from "@/graphql/schema/constants";
import definition from "@/graphql/schema/definitions/User";
import { Account } from "@/graphql/schema/generated/models/account.model";
import { Calendar } from "@/graphql/schema/generated/models/calendar.model";
import { CalendarEvent } from "@/graphql/schema/generated/models/calendarEvent.model";
import { Goal } from "@/graphql/schema/generated/models/goal.model";
import { Habit } from "@/graphql/schema/generated/models/habit.model";
import { Mantra } from "@/graphql/schema/generated/models/mantra.model";
import { Notebook } from "@/graphql/schema/generated/models/notebook.model";
import { Task } from "@/graphql/schema/generated/models/task.model";
import * as Scalars from "@/graphql/schema/scalars";
import * as Types from "@/graphql/schema/types";
import { Model } from "@/graphql/schema/types";
import { getModelForClass, ModelOptions, post, pre, prop as Property } from "@typegoose/typegoose";
import * as TypeGraphQL from "type-graphql-v2-fork";

@TypeGraphQL.ObjectType()
@ModelOptions(DEFAULT_MODEL_OPTIONS)
@pre<User>(
  "save",
  (definition?.hooks?.save?.pre as any) ??
    async function (next) {
      return next();
    }
)
@post<User>(
  "save",
  definition?.hooks?.save?.post ??
    (async (_instance) => {
      return;
    })
)
export class User extends Model {
  @TypeGraphQL.Field(() => Scalars.String, { nullable: true })
  @Property({ type: () => String, required: false, default: null })
  name?: Types.String | null;

  @TypeGraphQL.Field(() => Scalars.String, { nullable: false })
  @Property({ type: () => String, required: true, unique: true })
  email!: Types.String;

  @TypeGraphQL.Field(() => Scalars.Boolean, { nullable: true })
  @Property({ type: () => Boolean, required: false, default: null })
  emailVerified?: Types.Boolean | null;

  @TypeGraphQL.Field(() => Scalars.String, { nullable: true })
  @Property({ type: () => String, required: false, default: null })
  image?: Types.String | null;

  @TypeGraphQL.Field(() => Scalars.Boolean, { nullable: false })
  @Property({ type: () => Boolean, required: true, default: false })
  isAdmin!: Types.Boolean;

  @TypeGraphQL.Field(() => Scalars.JSON, { nullable: false })
  @Property({ required: true, default: {} })
  settings!: Types.UserSettings;

  @TypeGraphQL.Field(() => Scalars.DateTime, { nullable: true })
  @Property({ required: false, default: null })
  lastLogin?: Types.DateTime | null;

  @TypeGraphQL.Field(() => Scalars.String, { nullable: true })
  @Property({ type: () => String, required: false, select: false, default: null })
  password?: Types.String | null;

  @TypeGraphQL.Field(() => [Account], { nullable: true })
  @Property({ required: false, default: [] })
  accounts?: Account[] | null;

  @TypeGraphQL.Field(() => [Calendar], { nullable: true })
  @Property({ required: false, default: [] })
  calendars?: Calendar[] | null;

  @TypeGraphQL.Field(() => [CalendarEvent], { nullable: true })
  @Property({ required: false, default: [] })
  calendarEvents?: CalendarEvent[] | null;

  @TypeGraphQL.Field(() => [Goal], { nullable: true })
  @Property({ required: false, default: [] })
  goals?: Goal[] | null;

  @TypeGraphQL.Field(() => [Habit], { nullable: true })
  @Property({ required: false, default: [] })
  habits?: Habit[] | null;

  @TypeGraphQL.Field(() => [Mantra], { nullable: true })
  @Property({ required: false, default: [] })
  mantras?: Mantra[] | null;

  @TypeGraphQL.Field(() => [Notebook], { nullable: true })
  @Property({ required: false, default: [] })
  notebooks?: Notebook[] | null;

  @TypeGraphQL.Field(() => [Task], { nullable: true })
  @Property({ required: false, default: [] })
  tasks?: Task[] | null;
}

const UserModel = getModelForClass<typeof User>(User);
export default UserModel;